🧹 Consolidate scene argument validation logic
🎯 **What:** Extracted the `validateSceneArgs` function from `src/tools/composite/scenes.ts` into a shared helper in `src/tools/helpers/scene-parser.ts`.
💡 **Why:** This makes `validateSceneArgs` available for reuse across other scene-related tools, resolving the code health issue regarding argument validation duplication without modifying the tested logic inside `handleScenes`.
✅ **Verification:** Ran `pnpm test` (all 403 tests passed) and `pnpm run check` (all 70 files verified clean via Biome). Confirmed the `handleScenes` function successfully imports and utilizes the extracted validation helper.
✨ **Result:** Improved modularity and enabled reuse of consistent error formatting for scene-related tool inputs across the codebase.
